Addition of the "omissible" attribute for the @facets annotation, which indicates which facet can be "omitted" in the GAML statements (i.e. the "default" first facet).
Allows to write "do: write with: message::'something' " instead of "do action: write with: message::'something'", for instance.

This step was necessary for what is going to come soon : the ":" after the keyword will become optional, statements with only two facets (i.e. the "omissible" one and an other) will be able to drop the second one too, and actions/primitives with only one argument will accept simply the value (i.e. without the name of the argument), so that it will become possible to write things like :

"do write 'something' " instead of the previous syntax. 

git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@3615 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a