Revamped the code base of the processor to make it clearer and less cluttered.
GamaProcessor is now basically a (complicated) serializer of annotations and related code elements (classes, methods), which relies on a String format with separators. Documentation is now serialized together with the annotations.

JavaWriter is (at its name implies) a class that creates source code (GamlAdditions.java) based on the serialization result. 

A double, triple, even quadruple "clean" will be necessary !

git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@4394 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a