Important commit that changes two fundamental underpinnings of the meta-model and kernel of GAMA. Built-in species are now compiled only once (incl. world) and all the "world species" are subclasses of the built-in world, all the "experiment species" are subclasses of "experiment", etc.
The "world" and the simulation are now one and the same agent, instance of "model".

The experiments are now full species and the "experiment agent" can be provided with attributes and behaviors of its own.

Please be aware that this commit (temporarily) BREAKS Batch and Headless experiments. Their underlying classes are NOT functional (and do not compile).

git-svn-id: https://gama-platform.googlecode.com/svn/branches/GAMA_CURRENT@6387 b91ef52e-2b48-11df-9ab6-b1cdc0d4a89a