Fixes #1076. Addresses #1352.
Geometries can now be saved in shape files (not only agents). The recent
commit to allow saving agents in CSV files has been reworked to access
the attributes of the agents (and not only their extra attributes),
remove uninteresting attributes (like host, members, peers, etc.) and
output the CSV with a default `;` separator.