Fixes #1580. Fixes #1569
Refines the "experimental" Java2D display (and introduces another one,
called "canvas", that uses a different (and even more efficient)
technique to render directly on screen. Both are in their early stage
and should not be used beyond testing.
Addresses most of #1326 (only for Java2D displays right now).
Adds a preference for showing FPS in the overlay.