Addresses #1665
Pushes the scope towards more parts of GAMA to phase out the use of the
global scheduler.
The scope can now tell whether it is paused or not (either globally or
on user hold). Simulations can also know if they are on user hold, as
well as experiments.
Simplifies the API of the action executer (which now executes
IExecutable).

Signed-off-by: Alexis Drogoul <alexis.drogoul@gmail.com>