Fixes #1924. Fixes #1923
Changes the API of event / user_command : the location of the mouse is
not passed as an argument anymore. Instead, it is now always available
through the use of the #user_location unit. Event statements now accept
blocks in their action: facet, in addition to the name of an action. 

Signed-off-by: AlexisDrogoul <alexis.drogoul@gmail.com>