Further improvements to the validation/build process.
Allows the visit of IDescriptions to be pruned.
Changes the notification of resources to editors (now goes through the
indexer for better decoupling and tracking of modifications)
Reinstalls the documentation of constants (broken in the last commit)
Deploys the graph-based index to headless runs as well.
Prevents ISyntacticStatements to be loaded too soon in the parsing
process
Correctly waits until the indexer has finished its first run, before
parsing and validating resources.
Adds a notification button in the editor before the first validation
Correctly handles the addition/removal/changes in resources/projects to
trigger indexation

Signed-off-by: AlexisDrogoul <alexis.drogoul@gmail.com>