fixes #2225
However, requires a lot of testing. To use it:
grid cell files: [file0, file1, file2];

the first file (here file0) is used as reference for the grid width,
heights and location. All the values of the files are casted into float
value and are accessible through the bands built-in variables:

ask cell {
    write "value corresponding to file0:" + bands[0];
    write "value corresponding to file1:" + bands[1];
    write "value corresponding to file2:" + bands[2];
}