Implements a benchmarking infrastructure in GAMA (see #2462).
Accessed by adding `benchmark: true` to any experiment. Should have
almost no effect on the performances when not running, and a
very little one when running (a few kilobytes of memory and a little
decrease in speed).
Should be tested more carefully, though, especially for multi-thread,
multi-simulation scenarios, as well as batch experiments.