Improves the handling of files outside the workspace
In addition to the fix to #2456, add a number of enhancements:

- absolute files are now linked from the workspace automatically (in an
"external" folder inside the project) when they are detected in a model
file.

- the navigator correctly shows them as outside links, and they can be
manipulated like any other resource.

- the contents of GAMA files in the navigator now reflects the status of
the "used" files: local or in the workspace.

- the popup and hyperlink work for both types of files in the editor (if
they can be provided with metadata, of course). I.e. an image file
outside the workspace will display its size, type, etc.

- editors (shapefiles, osm, etc.) work for both types of files.

- linked files can reveal their target using the same command ("Reveal
in OS").

- files present in the "Uses" section of gama files in the navigator now
display the path of their target and can also reveal it in the
workspace.