Batch of commits related to files
1) Added preferences to control http connections (timeouts, empty cache,
retries, etc.)
2) Prevents a silent error in opengl displays when a texture cannot be
found
3) Extends the navigator to account for web resources used in models
(like outside files), with a built-in cache and linking mechanism
4) Strengthen the image loading (from file / web) to avoid double
loading and decoding
5) Fixes a long-standing bug in ImageLayerStatement (where two
simulations running in parallel could potentially change the other's
image) and avoids the double creation of files
6) Improve the creation of files (local, external or web) and compute
the absolute path (to either the file or the cached file) immediately.
localPath is now final, preventing any corruption from other objects.
7) Fixes a long-standing bug when printing (writing) list of files (due
to the serialization invoked outside of the scope)
8) Improves (a tiny bit) the loading time of GAMA by not checking the
configuration
9) Automatically detect when an absolute file is already part of the
workspace (and link to it instead of creating an intermediary link).