Fixes #2493 by organizing the preferences in two groups:
The ones usable in models (in both headless and GUI experiments), and
the ones belonging to the working environment. The Preferences.gaml file
has been restructured as well to show only the first ones. The
preferences view only displays the ids of the first ones and indicates
clearly that the second ones cannot be used in GAML.