Ugly fix for a problem for 3D display with opengl
The problem with basic shape (sphere, cube,...) was that with the
current version of version, the objects was not centered around their
location, and more specifically the z position : for instance, if you
write for an agent located at {0,0,0} the statement draw sphere(5.0);
the opengl display puts the agents on the "top of the table" (i.e. the
centered of the display sphere is {0,0,2.5}). It was not the case
before. So I added an ugly translation of the _center for sphere, cube,
pyramid and cylinder... This commit needs to be checked by an opengl
expert.