Deep reorganization of the GamaList, IList and GamaMap classes.
The two factories (GamaListFactory and GamaMapFactory) are now mandatory
to use and they produce IList and IMap (new interface) objects. Some of
these objects can be GamaList and GamaMap, as before, but also wrappers
around existing Collections and Arrays (for IList) or Maps (for IMap).
Wrappers allow to reduce the number of copies of collections, especially
for read-only attributes, like the keys and values of maps. Other types
of lists and maps may be introduced in the future, e.g. allowing to
switch to implementations based on ConcurrentHashMaps and concurrent
Lists when species schedule their agents in parallel.

Tens of models have been tested so as to ensure that the dynamics would
remain the same and that no bugs have been introduced. In case you
encounter one, please raise an issue immediately at
https://github.com/gama-platform/gama/issues . Compilation problems
might also occur in plugins still using GamaList and GamaMap directly.