Temporary workaround for issue #3068
The problem comes from the use of TopologicalOrderIterator of Jgrapht to
order the vertices of the graph according to their dependency (edges of
the graph) and that seems to forget some vertices.

The temporary fix consists in adding the forgotten elements at the end
of the list. A better solution (fixing TopologicalOrderIterator) should
be investigated.