Notify the population listeners after their initialization
Notify the population listeners of the creation of agents only after
they are completely initialized (incl. running the optional
initialization sequence). Correctly fires listeners when grid agents are
created.

default methods (corresponding to the the previous signature without the
RemoteSequence to execute) have been added to IPopulation so as to not
break compatibility. They call the new ones by passing null.