First addition of the JBox2D physics engine alongside Bullet (3D).
Not tested very thoroughly -- and not testable yet from GAML.The same
interfaces now allow to choose between JBullet (java), JBox2D (java),
Bullet (native). The goal, ultimately, is to have two choices : 2D or
3D, which would lead, by default, to either Box2D or Bullet (both
natives). But, in case the natives wouldn't work (or be loaded), to be
able to rely on the two corresponding java libraries.

More integration needs to be made with other plugins too. For the
moment, the two skills in physics only listen to "location" (i.e. to
detect a change of location coming from elsewhere), but it is
technically possigle to listen to any variable (change of shape, change
of speed, change of heading,...).