Allows (finally!) experiments to declare their own parameters
Thus eliminating the need to declare in global some attributes only
dedicated to display/outputs, for instance (path of icons, font to use,
color of the agents, gradients, etc.). These parameters can now be
safely declared and managed in the experiment itself, without
"polluting" the model.
The lookup remains the same: the variables used as attributes are first
searched in the model, then in the experiment (so as to keep the same
dynamics: seed, for instance, is still looked in the experiment).
Reserved for the moment to GUI experiments only.
Next step is to visually distinguish them (maybe also in a different
category) ?

Example:
```
experiment e {
	int a <- 100 parameter: "A";
	int b <- 1000;
	parameter "B" var: b;
}

will now display these two parameters, alongside the ones declared in
the model or declared targeting model's variables.