Several bits of additions (see warning below):
- FLAGS class to host flags used in the code (might be loaded through
properties/ini file at some point)
- Possibility to choose to use the old FPS Animator in OpenGL (see
#3164). FLAGS.USE_OLD_ANIMATOR
- Addition of a preference to limit or not the FPS (GamaPreferences)
- Possibility for Preferences to use the configuration area of
applications, i.e. to be different for each instance of GAMA. Controlled
for the moment by FLAGS.USE_JRE_PREFERENCE_STORE (true corresponds to
the default behavior).
- Addition of a generic GamaPreferenceStore that can use both stores
- Addition of a button to export preferences as a *.prefs file (standard
Eclipse format)
- **Warning**: preferences might be reset with this version. If you need
to reinstall them, save them first in the previous version of GAMA
(either as a gama.prefs or a Preferences.gaml file) then choose to
reload them in the new version of GAMA (in the UI for gama.prefs, by
running the model for Preferences.gaml).