Allows preferences to be overriden if passed as VM arguments to GAMA
e.g. -Dpref_use_pooling=true in the VM arguments of the run
configuration or gama.ini file will override whatever value is defined
in the preferences.
**WARNING**: It has the side effect of changing the value of the
preference for all instances if -Duse_global_preference_store=false is
not set at the same time (or if the other instances do not defined their
own override).