Possible fix for #3268.
Please test as the behavior of imports is slightly changed : now in case
of multiple imports of the same model, the deepest ones in the
dependency graph are loaded before the others, correctly respecting the
hierarchy of imports. Same for the order in which the imports are
declared, which now has a semantic value : the models will be loaded in
the order they are declared.