Fixes #3299
The fill/line properties of glPolygonMode seems to be altered outside of
the drawings done by GAMA. Fixing it means changing the mode quite
often, which might have an impact on large drawings. For mesh layers,
the cause was that the change of property at the display level was not
reflected in the MeshDrawer : the buffer containing line colors was
never correctly initialized.