Preliminary refactoring of the Converter hierarchy (#3428). PLEASE TEST
Introduction of IGamaConverter and AbstractGamaConverter, to prepare a
real factory. Refactoring of all the affected classes.
Reduces the garbage produced by keeping only one instance of each
converter and passing it the current scope. THIS MUST BE TESTED SO AS TO
USE, MAYBE, A THREADLOCAL VERSION IN CASE OF CONCURRENCY PROBLEMS