Provides a possible solution to #3428. **NEEDS TO BE TESTED**
Introduction of the @converter annotation (in the serialize plugin),
which allows to declare the converter as an annotation to the IType. The
converters are then discovered and instantiated, based on the class
wrapped by the IType, by the Converters class in the serialize plugin.

Allows to mix hardcoded (for all the built-in types in the core plugin)
and dynamically defined converters.