Addresses #3472, #3468 and marginally #3473
- #3468: Two new preferences allow to list required plugins in model
files when editing them, and to check their existence if they are
listed. Still experimental as the moment, especially because of
"display" plugins which might not be present in headless operations
- #3472: Adds a functional 'continue' statement and rewrites part of the
implementation of 'break', adding a FlowStatus to IScope. Requires to be
better tested, though ! Annotation processor had to be recompiled to
take into account breakable/continuable statements. Addition of a new
model in the library dedicated to break / continue.
- #3473: Adds functions in GamaBundleLoader to retrieve and list the
features installed at runtime