Addresses #3498.
A tricky bug, not easy to track and which supposed to change some logic
in the management of the layers. Hopefully the fix doesnt break other
mechanisms ! Please test it thoroughly.
In addition, the layers menu now indicates the current transparency of a
layer and disables aspect/transparency options for non-refreshable
layers.