Provides a number of changes in random number generation.
1) Removal of the "cellular" algorithm, barely used, too simple, and not
thread safe.
2) Reorganisation of the classes for better readability and
compatibility with Java RandomGenerator.
3) Addition of two thread safe algorithms : `parallel`, a mersenne
twister with a reentrant lock to prevent concurrent stage change -- it
remains deterministic at the expense of implicitly synchronising all the
threads accessing it (at least the ones accessing the critical section)
; `threaded`, a RNG based on `ThreadLocalRandom`, very fast even in
highly concurrent, but that cannot be seedable (and therefore is not
deterministic). The first one is probably a good choice for parallel
simulations that try to keep some determinism; the second one for
interactive parallel simulations.
4) Better documentation for these algorithms in GAML.