Ability to lock any Zoomable display.
This commits adds the toggleLock() method to the Zoomable interface.
When a Zoomable item is locked, zoomIn(), zoomOut() and dragging of the
view pane have no effect, but zoomFit() is still active.

The ability to lock a display was already available for OpenGL displays,
thanks to the Camera interface. The feature has been generalised, and is
now available for:
- Java2DDisplaySurface
- GISFileViewer
- ImageViewer

The already existing lock control has been added to toolbars
accordingly, to any Zoomable display (see ZoomController).