Implements mouse_drag events handling.
Adds a mouse_drag event to GAML. The mouse_drag event is the kind of events
emitted when the mouse is moved while the mouse button is down (while mouse_move
events are emitted when the mouse button is up). This addition adds significant
abilities to GAML, since mouse_move events (or any other kind of event) was not
sent while the mouse button was down (at least on linux platforms), what
prevented to implement "drag and drop" like features within GAMA models.

- ummisco.gama.annotations: The mouse_drag keyword has been added to GAML
- msi.gama.core: EventLayer, EventLayerStatement, IEventLayerListener and
  MouseEventLayerDelegate have been adapted to handle mouse dragged events.
- msi.gama.models: A Mouse Drag GAMA model has been added to the library models
  in the User Interaction section to demonstrate how the feature can be used to
  drag and drop an agent in both OpenGL and Java2D displays.
- ummisco.gama.ui.experiment: The LayeredDisplayMultiListener now properly
  handles drag events
- ummisco.gama.java2D: The Java2DDisplaySurface now dispatches SWT.DragDetect
  events
- ummisco.gama.opengl: The SWTOpenGLDisplaySurface now dispatches SWT.DragDetect
  events

- The feature is still not properly supported on tablets, but this hopefully
  will be fixed in next commits.