Solves a mouse_down issue, critical on touch screens
In OpenGL displays, the mouse_down events was not properly handled since
the regular SWT event layer triggered the GAML mouse_down events, and
only after that (from an other SWT event) the OpenGL CameraHelper
updated the world position, i.e. the user_location. In consequence, the
user_location available from the mouse_down GAML action was not up to
date. This was not critical when using a mouse, since in this case
mouse_down events are always preceded by mouse_move events that also
update the user_location. However, this was critical on touch screens,
where independent mouse_down events are triggered from random places
without mouse_move events.

In order to solve the issue, the world position in OpenGL displays is
now updated from the SWTOpenGLDisplaySurface, from the same SWT mouse
down events that trigger corresponding GAML actions, and not from the
CameraHelper. However, the CameraHelper is still free to use the updated
world position, even if it also manages its own and independent
mousePosition.