Various enhancements
1) Reorganizes the SaveStatement implementation using the same design
pattern than CreateStatement. Allows the definition of ISaveDelegate
implementations in plugins (for saving specific data or in specific
formats). Refactoring of the different "savers" into a hierarchy of
ISaveDelegates and makes them all stateless. Redeclares the existing
"savers" as extensions implementations in plugin.xml of msi.gama.core.

2) Corrects two problems with icons (no update of the Application,
simplification of the loading of icons at startup.