Changes in AbstractAgent and GamaShape (#3847)
- Let AbstractAgent declare and use agents' attributes, separated from
GamaShape. No impact on the API, but it might have some on the usage,
specifically in the area of GIS attributes. For instance, now, GIS
attributes loaded from the file will not replace the values of agents
attributes with the same names ==> needs to be tested in large GIS-based
models. Necessary first step for addressing #3847.
- Introduces GamaShapeFactory to create GamaShapes more safely, and
deprecates all GamaShape constructors. No API change for the moment (but
it will come !)
- Adapts part of the dependencies / code to Eclipse 2023-09 (might break
local compilation)
- Some various simple changes (comments, etc.)